retrieving data from FB is not working
