+++ src/index.js
@@ -11,9 +11,21 @@ import App from './components/App';
 import DetailedPin from './components/DetailedPin';
 import AddingRecommendation from './containers/AddingRecommendation';
 import * as actions from './actions/auth';
-
+import firebase from 'firebase';
 // const middleware = [thunk];
 require('!style!css!sass!./styles/main.scss');
+
+const firebaseApp = {
+  apiKey: 'AIzaSyDwnsJcOAy2cP2lyTFdWcanOMTf2DRZpIs',
+  authDomain: 'local-cb336.firebaseapp.com',
+  databaseURL: 'https://local-cb336.firebaseio.com',
+  storageBucket: 'local-cb336.appspot.com',
+  messagingSenderId: '514582937003'
+};
+
+console.log('WHAAAT');
+
+firebase.initializeApp(firebaseApp);
 
 const enhancers = compose(
   applyMiddleware(thunk),
